OneDrive - MUET/Desktop/SAAS1/seo-engine/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
httpx>=0.25.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
sentry-sdk[fastapi]>=1.39.0
redis>=5.0.0
